Photoshop - Pixlr UFO area 51 edits
A friend MK7Carbon asked me to put together an UFO / area 51 edit with his car for Instagram. This is what i was able to whip up for him. I have included some edits.

I had some shading/ blending issues, so i tried to hide it with editing the picture.